pyproject.toml
[tool.poetry]
name = "project-template"
version = "1.0.0"
description = ""
authors = ["Paulo Roberto Urio <paulourio@gmail.com>"]
readme = "README.md"
packages = [{ include = "project", from = "source" }]

[tool.poetry.dependencies]
python = ">=3.8, <3.9"
dynaconf = "^3.1.8"

[tool.poetry.group.modeling.dependencies]
catboost = "^1.0.6"
numpy = "^1.22.4"
pandas = "^1.4.2"
scikit-learn = "^1.1.1"

[tool.poetry.group.evaluation.dependencies]
hyperopt = "^0.2.7"
shap = "^0.40.0"
statsmodels = "^0.13.2"

[tool.poetry.group.visualization.dependencies]
altair = "^4.2.0"
matplotlib = "^3.5.2"
seaborn = "^0.11.2"

[tool.poetry.group.formatters.dependencies]
humanfriendly = "^10.0"
Jinja2 = "^3.1.2"
parsimonious = "^0.9.0"
prettydiff = {extras = ["terminal"], version = "^0.1.0"}
prettyprinter = "^0.18.0"
Pygments = "^2.12.0"
tabulate = "^0.8.9"

[tool.poetry.group.file_formats.dependencies]
joblib = "^1.1.0"
pyarrow = "<8.0.0"
PyYAML = "^6.0"

[tool.poetry.group.google.dependencies]
apache-beam = {version = "2.39.0", extras = ["gcp"]}
gcsfs = "^2022.5.0"
google-cloud-aiplatform = "<1.13.1"
google-cloud-bigquery = "<3.1.0"
google-cloud-logging = "^3.1.1"
google-cloud-storage = "<2.3.0"

[tool.poetry.group.dev.dependencies]
autopep8 = "^1.6.0"
coloredlogs = "^15.0.1"
ipykernel = "^6.13.0"
python-dotenv = "^0.20.0"

[tool.poetry.group.ci.dependencies]
apache-beam = {version = "2.39.0", extras = ["test"]}
flake8 = "^4.0.1"
hypothesis = "^6.46.11"
mypy = "^0.960"
numpydoc = "^1.3.1"
pycodestyle = "^2.8.0"
pydocstyle = "^6.1.1"
pylint = "^2.14.0"
pytest = "<5.0"
pytest-cov = "^3.0.0"
types-dataclasses = "^0.6.5"
types-Markdown = "^3.3.27"
types-mock = "^4.0.14"
types-parsimonious = "^0.8.5"
types-python-dateutil = "^2.8.17"
types-pytz = "^2021.3.8"
types-PyYAML = "^6.0.8"
types-regex = "^2021.11.10.5"
types-tabulate = "^0.8.9"
types-toposort = "^1.7.3"

[tool.poetry.group.debugging.dependencies]
memory-profiler = "^0.60.0"
Pympler = "^1.0.1"

[tool.mypy]
plugins = "numpy.typing.mypy_plugin"

[tool.pydocstyle]
convention = 'numpy'

[tool.pytest.ini_options]
markers = [
    "slow",
]
log_level = "DEBUG"

[build-system]
requires = ["poetry-core>=1.2.0"]
build-backend = "poetry.core.masonry.api"
